requirements.txt
python-dotenv
langchain-openai
langchain-ollama
anthropic
groq
